fixed tick type

+++ Firmcheese/src/key_event.hpp
@@ -11,11 +11,11 @@ enum class EKeyEvent {
 using keycode_t = uint16_t;
 
 struct KeyEvent {
-  KeyEvent(keycode_t code = 0, EKeyEvent event = EKeyEvent::None, int16_t tick_ms = 0)
+  KeyEvent(keycode_t code = 0, EKeyEvent event = EKeyEvent::None, uint32_t tick_ms = 0)
       : code_(code), event_(event), tick_ms_(tick_ms) {}
   keycode_t code_;
   EKeyEvent event_;
-  int16_t tick_ms_;
+  uint32_t tick_ms_;
 
   inline bool isPressed() const { return event_ == EKeyEvent::Pressed; }
 };
